+++ src/main/java/com/goalracha/domain/Member.java
@@ -56,7 +56,7 @@ public class Member {
     @Enumerated(EnumType.STRING)
     @Column(name = "type")
     private MemberRole type; //유저타입 user,owner,admin
-    private String password;
+
 
     public void joinMember(Long uNo, String name, String nickname, String tel) {
         this.uNo = uNo;
@@ -67,6 +67,6 @@ public class Member {
     }
 
     public void setPassword(String password) {
-        this.password = password;
+        this.pw = password;
     }
 }
